- renamed middleware to httpMultipartBodyParser - Added docs - Added support for specifying busboy options - Removed double export (exporting only middleware for consistency)
