+++ src/middlewares/index.js
@@ -6,10 +6,10 @@ module.exports = {
   httpErrorHandler: require('./httpErrorHandler'),
   httpEventNormalizer: require('./httpEventNormalizer'),
   httpHeaderNormalizer: require('./httpHeaderNormalizer'),
+  httpMultipartBodyParser: require('./httpMultipartBodyParser'),
   httpPartialResponse: require('./httpPartialResponse'),
   httpSecurityHeaders: require('./httpSecurityHeaders'),
   jsonBodyParser: require('./jsonBodyParser'),
-  multipartFormDataParser: require('./multipartFormDataParser').middleware,
   s3KeyNormalizer: require('./s3KeyNormalizer'),
   secretsManager: require('./secretsManager'),
   ssm: require('./ssm'),
